+++ new_version/src/main/java/pl/krzysiek/services/ReadyMealService.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface ReadyMealService {
 
-    List<ReadyMealDetails> converterDataFromForm(Integer[] gramsPortion, Integer[] food_ingredient_id);
+    List<ReadyMealDetails> converterDataFromForm(Double[] gramsPortion, Integer[] food_ingredient_id);
     Boolean saveReadyMeal(ReadyMeal readyMeal);
     List<ReadyMeal> findAlle();
